requirements.txt
faiss_cpu==1.7.4
fastapi==0.124.4
numpy==1.24.4
onnxruntime==1.23.2
opencv_python==4.8.0.76
pydantic==2.12.5
starlette==0.50.0
